Get StackOverflow Questions For A List of Tags

Extract StackOverflow questions for a given set of tags, exluding questions related to an optional list of tags. This code is based on the tidyverse tweets bot by Jake Thompson.

get_stack_questions(extracted_tags = NULL, excluded_tags = NULL,
  time_window = 60, add_process_fn = NULL)

Arguments

extracted_tags
A character vector of tags to extract questions for.

excluded_tags
A character vector of tags to exclude questions from the output.

time_window
Numeric, the timewindow to extract questions for (in minutes. Defaults to 60 minutes. Note that a maximum of 100 questions are extracted by the StackOverflow api, if you expect your tags to have a high volume of questions reduce this. If using this function as part of a sheduled bot makes sure this time frame corresponds the frequency with which your bot is running.

add_process_fn
A function. If additional munging or data extraction is required add to a custom function and pass here. This function must except a dataframe as its first arguement and return a dataframe. See the examples for details.

Value

A list containing the title, creation date, link, and tags for StackOverflow questions related to your selected tags.

Examples

## Basic call
get_stack_questions(extracted_tags = "ggplot2",
                    excluded_tags = "plotly",
                    time_window = 60*24)

#> $title
#>  [1] "Cumulative plot in ggplot2"                                                                   
#>  [2] "Misaligned points in boxplots using ggplot2"                                                  
#>  [3] "How to increase resolution of ggplots when exporting using officer R"                         
#>  [4] "Showing X axis value label on top of the histogram"                                           
#>  [5] "unable to start png() device for plot names create with \"paste\" function"                   
#>  [6] "ggplot add horizontal line inside for loop"                                                   
#>  [7] "R Shiny not plotting some ggplot2 layers"                                                     
#>  [8] "Days for major and hours for minor ticks"                                                     
#>  [9] "ggplot2 - add proportion and count values to barchart"                                        
#> [10] "How to draw the background and change the color of each laboratory?"                          
#> [11] "Can we change angle of text with geom_label in ggplot2?"                                      
#> [12] "Unexpected Walker alias table output using ggplot2"                                           
#> [13] "Showing a legend on a multiple-Y axis plot in ggplot"                                         
#> [14] "Create multiple 2-column bar plots plotting a particular factor against all factors in a loop"
#> [15] "ggplot alpha = 0 not working"                                                                 
#> [16] "Need smoothed line on histogram. Why does geom_density draw two lines?"                       
#> [17] "How to use multiple columns as the frame of an animation in gganimate package (R)?"           
#> [18] "Plotting a table from dataframe values using both ggplot2 and gridExtra packages in R"        
#> [19] "Is it possible to save a ggplot in RDS format with a global variable in the ggplot() call?"   
#> [20] "Difference in ggplot in RMD and flexdashboard"                                                
#> [21] "Mark in ggplot when the observations go below a certain number"                               
#> [22] "choices combination,order & tree"                                                             
#> [23] "what's difference between \"method\" argument values in geom_smooth()"                        
#> [24] "ggplot2 is not plotting two groups"                                                           
#> 
#> $creation_date
#>  [1] "2018-03-12 15:27:30 UTC" "2018-03-12 16:35:16 UTC"
#>  [3] "2018-03-12 17:15:00 UTC" "2018-03-12 17:44:54 UTC"
#>  [5] "2018-03-12 19:33:23 UTC" "2018-03-12 20:38:53 UTC"
#>  [7] "2018-03-12 21:47:59 UTC" "2018-03-12 23:43:57 UTC"
#>  [9] "2018-03-13 00:32:04 UTC" "2018-03-13 00:41:32 UTC"
#> [11] "2018-03-13 00:55:20 UTC" "2018-03-13 01:07:19 UTC"
#> [13] "2018-03-13 01:41:17 UTC" "2018-03-13 02:06:09 UTC"
#> [15] "2018-03-13 02:33:14 UTC" "2018-03-13 02:41:51 UTC"
#> [17] "2018-03-13 02:56:06 UTC" "2018-03-13 07:13:31 UTC"
#> [19] "2018-03-13 08:57:10 UTC" "2018-03-13 11:48:21 UTC"
#> [21] "2018-03-13 12:20:07 UTC" "2018-03-13 13:19:15 UTC"
#> [23] "2018-03-13 13:42:56 UTC" "2018-03-13 14:35:21 UTC"
#> 
#> $link
#>  [1] "https://stackoverflow.com/questions/49238865/cumulative-plot-in-ggplot2"                                                      
#>  [2] "https://stackoverflow.com/questions/49240194/misaligned-points-in-boxplots-using-ggplot2"                                     
#>  [3] "https://stackoverflow.com/questions/49240961/how-to-increase-resolution-of-ggplots-when-exporting-using-officer-r"            
#>  [4] "https://stackoverflow.com/questions/49241460/showing-x-axis-value-label-on-top-of-the-histogram"                              
#>  [5] "https://stackoverflow.com/questions/49243161/unable-to-start-png-device-for-plot-names-create-with-paste-function"            
#>  [6] "https://stackoverflow.com/questions/49244049/ggplot-add-horizontal-line-inside-for-loop"                                      
#>  [7] "https://stackoverflow.com/questions/49244946/r-shiny-not-plotting-some-ggplot2-layers"                                        
#>  [8] "https://stackoverflow.com/questions/49246106/days-for-major-and-hours-for-minor-ticks"                                        
#>  [9] "https://stackoverflow.com/questions/49246486/ggplot2-add-proportion-and-count-values-to-barchart"                             
#> [10] "https://stackoverflow.com/questions/49246553/how-to-draw-the-background-and-change-the-color-of-each-laboratory"              
#> [11] "https://stackoverflow.com/questions/49246650/can-we-change-angle-of-text-with-geom-label-in-ggplot2"                          
#> [12] "https://stackoverflow.com/questions/49246760/unexpected-walker-alias-table-output-using-ggplot2"                              
#> [13] "https://stackoverflow.com/questions/49247040/showing-a-legend-on-a-multiple-y-axis-plot-in-ggplot"                            
#> [14] "https://stackoverflow.com/questions/49247222/create-multiple-2-column-bar-plots-plotting-a-particular-factor-against-all-fact"
#> [15] "https://stackoverflow.com/questions/49247417/ggplot-alpha-0-not-working"                                                      
#> [16] "https://stackoverflow.com/questions/49247483/need-smoothed-line-on-histogram-why-does-geom-density-draw-two-lines"            
#> [17] "https://stackoverflow.com/questions/49247591/how-to-use-multiple-columns-as-the-frame-of-an-animation-in-gganimate-package-r" 
#> [18] "https://stackoverflow.com/questions/49250257/plotting-a-table-from-dataframe-values-using-both-ggplot2-and-gridextra-packages"
#> [19] "https://stackoverflow.com/questions/49251921/is-it-possible-to-save-a-ggplot-in-rds-format-with-a-global-variable-in-the-ggpl"
#> [20] "https://stackoverflow.com/questions/49255392/difference-in-ggplot-in-rmd-and-flexdashboard"                                   
#> [21] "https://stackoverflow.com/questions/49256057/mark-in-ggplot-when-the-observations-go-below-a-certain-number"                  
#> [22] "https://stackoverflow.com/questions/49257258/choices-combination-order-tree"                                                  
#> [23] "https://stackoverflow.com/questions/49257753/whats-difference-between-method-argument-values-in-geom-smooth"                  
#> [24] "https://stackoverflow.com/questions/49258931/ggplot2-is-not-plotting-two-groups"                                              
#> 
#> $tags
#>  [1] "r,plot,ggplot2,cumsum"                           
#>  [2] "r,ggplot2"                                       
#>  [3] "r,ggplot2,officer"                               
#>  [4] "r,ggplot2"                                       
#>  [5] "r,plot,ggplot2"                                  
#>  [6] "r,ggplot2"                                       
#>  [7] "r,ggplot2,shiny,visualization,data-visualization"
#>  [8] "r,ggplot2"                                       
#>  [9] "r,ggplot2"                                       
#> [10] "r,ggplot2,tidyverse"                             
#> [11] "r,ggplot2"                                       
#> [12] "r,ggplot2,stacked-chart"                         
#> [13] "r,ggplot2"                                       
#> [14] "r,ggplot2"                                       
#> [15] "r,ggplot2"                                       
#> [16] "r,ggplot2,histogram,smoothing,density-plot"      
#> [17] "r,animation,ggplot2,visualization,gganimate"     
#> [18] "r,ggplot2,table,gridextra"                       
#> [19] "r,ggplot2"                                       
#> [20] "r,ggplot2,flexdashboard"                         
#> [21] "r,ggplot2"                                       
#> [22] "r,ggplot2"                                       
#> [23] "r,ggplot2,smoothing"                             
#> [24] "ggplot2,tidyverse"                               
#> 

## Add additional processing step
library(dplyr)
add_process_fn <- function(df) {
      df %>%
      mutate(id = 1:n())
}

get_stack_questions(extracted_tags = "ggplot2",
                    excluded_tags = "plotly",
                    time_window = 60*24,
                    add_process_fn = add_process_fn)

#> $title
#>  [1] "Cumulative plot in ggplot2"                                                                   
#>  [2] "Misaligned points in boxplots using ggplot2"                                                  
#>  [3] "How to increase resolution of ggplots when exporting using officer R"                         
#>  [4] "Showing X axis value label on top of the histogram"                                           
#>  [5] "unable to start png() device for plot names create with \"paste\" function"                   
#>  [6] "ggplot add horizontal line inside for loop"                                                   
#>  [7] "R Shiny not plotting some ggplot2 layers"                                                     
#>  [8] "Days for major and hours for minor ticks"                                                     
#>  [9] "ggplot2 - add proportion and count values to barchart"                                        
#> [10] "How to draw the background and change the color of each laboratory?"                          
#> [11] "Can we change angle of text with geom_label in ggplot2?"                                      
#> [12] "Unexpected Walker alias table output using ggplot2"                                           
#> [13] "Showing a legend on a multiple-Y axis plot in ggplot"                                         
#> [14] "Create multiple 2-column bar plots plotting a particular factor against all factors in a loop"
#> [15] "ggplot alpha = 0 not working"                                                                 
#> [16] "Need smoothed line on histogram. Why does geom_density draw two lines?"                       
#> [17] "How to use multiple columns as the frame of an animation in gganimate package (R)?"           
#> [18] "Plotting a table from dataframe values using both ggplot2 and gridExtra packages in R"        
#> [19] "Is it possible to save a ggplot in RDS format with a global variable in the ggplot() call?"   
#> [20] "Difference in ggplot in RMD and flexdashboard"                                                
#> [21] "Mark in ggplot when the observations go below a certain number"                               
#> [22] "choices combination,order & tree"                                                             
#> [23] "what's difference between \"method\" argument values in geom_smooth()"                        
#> [24] "ggplot2 is not plotting two groups"                                                           
#> 
#> $creation_date
#>  [1] "2018-03-12 15:27:30 UTC" "2018-03-12 16:35:16 UTC"
#>  [3] "2018-03-12 17:15:00 UTC" "2018-03-12 17:44:54 UTC"
#>  [5] "2018-03-12 19:33:23 UTC" "2018-03-12 20:38:53 UTC"
#>  [7] "2018-03-12 21:47:59 UTC" "2018-03-12 23:43:57 UTC"
#>  [9] "2018-03-13 00:32:04 UTC" "2018-03-13 00:41:32 UTC"
#> [11] "2018-03-13 00:55:20 UTC" "2018-03-13 01:07:19 UTC"
#> [13] "2018-03-13 01:41:17 UTC" "2018-03-13 02:06:09 UTC"
#> [15] "2018-03-13 02:33:14 UTC" "2018-03-13 02:41:51 UTC"
#> [17] "2018-03-13 02:56:06 UTC" "2018-03-13 07:13:31 UTC"
#> [19] "2018-03-13 08:57:10 UTC" "2018-03-13 11:48:21 UTC"
#> [21] "2018-03-13 12:20:07 UTC" "2018-03-13 13:19:15 UTC"
#> [23] "2018-03-13 13:42:56 UTC" "2018-03-13 14:35:21 UTC"
#> 
#> $link
#>  [1] "https://stackoverflow.com/questions/49238865/cumulative-plot-in-ggplot2"                                                      
#>  [2] "https://stackoverflow.com/questions/49240194/misaligned-points-in-boxplots-using-ggplot2"                                     
#>  [3] "https://stackoverflow.com/questions/49240961/how-to-increase-resolution-of-ggplots-when-exporting-using-officer-r"            
#>  [4] "https://stackoverflow.com/questions/49241460/showing-x-axis-value-label-on-top-of-the-histogram"                              
#>  [5] "https://stackoverflow.com/questions/49243161/unable-to-start-png-device-for-plot-names-create-with-paste-function"            
#>  [6] "https://stackoverflow.com/questions/49244049/ggplot-add-horizontal-line-inside-for-loop"                                      
#>  [7] "https://stackoverflow.com/questions/49244946/r-shiny-not-plotting-some-ggplot2-layers"                                        
#>  [8] "https://stackoverflow.com/questions/49246106/days-for-major-and-hours-for-minor-ticks"                                        
#>  [9] "https://stackoverflow.com/questions/49246486/ggplot2-add-proportion-and-count-values-to-barchart"                             
#> [10] "https://stackoverflow.com/questions/49246553/how-to-draw-the-background-and-change-the-color-of-each-laboratory"              
#> [11] "https://stackoverflow.com/questions/49246650/can-we-change-angle-of-text-with-geom-label-in-ggplot2"                          
#> [12] "https://stackoverflow.com/questions/49246760/unexpected-walker-alias-table-output-using-ggplot2"                              
#> [13] "https://stackoverflow.com/questions/49247040/showing-a-legend-on-a-multiple-y-axis-plot-in-ggplot"                            
#> [14] "https://stackoverflow.com/questions/49247222/create-multiple-2-column-bar-plots-plotting-a-particular-factor-against-all-fact"
#> [15] "https://stackoverflow.com/questions/49247417/ggplot-alpha-0-not-working"                                                      
#> [16] "https://stackoverflow.com/questions/49247483/need-smoothed-line-on-histogram-why-does-geom-density-draw-two-lines"            
#> [17] "https://stackoverflow.com/questions/49247591/how-to-use-multiple-columns-as-the-frame-of-an-animation-in-gganimate-package-r" 
#> [18] "https://stackoverflow.com/questions/49250257/plotting-a-table-from-dataframe-values-using-both-ggplot2-and-gridextra-packages"
#> [19] "https://stackoverflow.com/questions/49251921/is-it-possible-to-save-a-ggplot-in-rds-format-with-a-global-variable-in-the-ggpl"
#> [20] "https://stackoverflow.com/questions/49255392/difference-in-ggplot-in-rmd-and-flexdashboard"                                   
#> [21] "https://stackoverflow.com/questions/49256057/mark-in-ggplot-when-the-observations-go-below-a-certain-number"                  
#> [22] "https://stackoverflow.com/questions/49257258/choices-combination-order-tree"                                                  
#> [23] "https://stackoverflow.com/questions/49257753/whats-difference-between-method-argument-values-in-geom-smooth"                  
#> [24] "https://stackoverflow.com/questions/49258931/ggplot2-is-not-plotting-two-groups"                                              
#> 
#> $tags
#>  [1] "r,plot,ggplot2,cumsum"                           
#>  [2] "r,ggplot2"                                       
#>  [3] "r,ggplot2,officer"                               
#>  [4] "r,ggplot2"                                       
#>  [5] "r,plot,ggplot2"                                  
#>  [6] "r,ggplot2"                                       
#>  [7] "r,ggplot2,shiny,visualization,data-visualization"
#>  [8] "r,ggplot2"                                       
#>  [9] "r,ggplot2"                                       
#> [10] "r,ggplot2,tidyverse"                             
#> [11] "r,ggplot2"                                       
#> [12] "r,ggplot2,stacked-chart"                         
#> [13] "r,ggplot2"                                       
#> [14] "r,ggplot2"                                       
#> [15] "r,ggplot2"                                       
#> [16] "r,ggplot2,histogram,smoothing,density-plot"      
#> [17] "r,animation,ggplot2,visualization,gganimate"     
#> [18] "r,ggplot2,table,gridextra"                       
#> [19] "r,ggplot2"                                       
#> [20] "r,ggplot2,flexdashboard"                         
#> [21] "r,ggplot2"                                       
#> [22] "r,ggplot2"                                       
#> [23] "r,ggplot2,smoothing"                             
#> [24] "ggplot2,tidyverse"                               
#> 
#> $id
#>  [1] 24 23 22 21 20 19 18 17 16 15 14 13 12 11 10  9  8  7  6  5  4  3  2  1
#>