Post A Tweet for Each Question

Format and post StackOverflow questions as tweets, optionally returning them as a formatted list. This code is based on the tidyverse tweets bot by Jake Thompson.

post_stack_tweets(questions = NULL, hashtags = "rstats", post = TRUE)

Arguments

questions
A list of questions to tweet as produced by get_stack_questions.

hashtags
A character vector of hashtags to attach to the tweet. See add_twitter_api for details.

post
Logical, defaults to TRUE. Should the formatted tweets be published on twitter, if FALSE then the tweets are instead returned as a list.

Value

Either nothing is returned or a formated list of tweets.

Examples

ggplot2_qs <-  get_stack_questions(extracted_tags = "ggplot2",
                                  excluded_tags = "plotly",
                                  time_window = 60*24)

post_stack_tweets(ggplot2_qs, hashtags = "rstats", post = FALSE)

#> [[1]]
#> Dynamically change background color of ggplot2 (r,ggplot2) #rstats https://stackoverflow.com/questions/49408969/dynamically-change-background-color-of-ggplot2
#> 
#> [[2]]
#> How to create multiple (6) plots with ggplot and save them to a pdf file? (r,pdf,ggplot2) #rstats https://stackoverflow.com/questions/49412035/how-to-create-multiple-6-plots-with-ggplot-and-save-them-to-a-pdf-file
#> 
#> [[3]]
#> How do I plot PostGIS geometries using geom_sf? (r,ggplot2,postgis,spatial) #rstats https://stackoverflow.com/questions/49413102/how-do-i-plot-postgis-geometries-using-geom-sf
#> 
#> [[4]]
#> adding legend to plot with 2 geom points (r,ggplot2) #rstats https://stackoverflow.com/questions/49415013/adding-legend-to-plot-with-2-geom-points
#> 
#> [[5]]
#> ggplot2: axis.line and panel.border have different width (r,plot,ggplot2) #rstats https://stackoverflow.com/questions/49415556/ggplot2-axis-line-and-panel-border-have-different-width
#> 
#> [[6]]
#> geom_text at certain position in a pie chart (r,ggplot2,tidyverse) #rstats https://stackoverflow.com/questions/49416386/geom-text-at-certain-position-in-a-pie-chart
#> 
#> [[7]]
#> Connected points in ggplot boxplot (r,ggplot2,data-visualization,boxplot) #rstats https://stackoverflow.com/questions/49417250/connected-points-in-ggplot-boxplot
#> 
#> [[8]]
#> R: loop through data.table and plot with ggplot2 (r,ggplot2) #rstats https://stackoverflow.com/questions/49417892/r-loop-through-data-table-and-plot-with-ggplot2
#> 
#> [[9]]
#> multivariate bar chart in R ggplot (r,ggplot2,bar-chart,data-visualization) #rstats https://stackoverflow.com/questions/49418057/multivariate-bar-chart-in-r-ggplot
#> 
#> [[10]]
#> Error in filter_impl: Evaluation error: Not compatible with requested type (r,ggplot2,shiny,dplyr) #rstats https://stackoverflow.com/questions/49418083/error-in-filter-impl-evaluation-error-not-compatible-with-requested-type
#> 
#> [[11]]
#> What is the best way to add ratio lines to plot in ggplot2? (r,plot,ggplot2,scatter-plot) #rstats https://stackoverflow.com/questions/49418545/what-is-the-best-way-to-add-ratio-lines-to-plot-in-ggplot2
#> 
#> [[12]]
#> Adding Regression Line Equation and R2 on SEPARATE LINES graph (r,ggplot2) #rstats https://stackoverflow.com/questions/49418552/adding-regression-line-equation-and-r2-on-separate-lines-graph
#> 
#> [[13]]
#> Change order of months scale_x_date ggplot (python,python-2.7,ggplot2,python-ggplot) #rstats https://stackoverflow.com/questions/49418669/change-order-of-months-scale-x-date-ggplot
#> 
#> [[14]]
#> ROC Curve in Shiny (r,ggplot2,shiny,roc,proc-r-package) #rstats https://stackoverflow.com/questions/49420225/roc-curve-in-shiny
#> 
#> [[15]]
#> Set color parameter in ggplot as optional - Shiny Dashboard (r,ggplot2,shinydashboard) #rstats https://stackoverflow.com/questions/49420853/set-color-parameter-in-ggplot-as-optional-shiny-dashboard
#> 
#> [[16]]
#> gridExtra - when I knit for pdf,i still get the big graphs (r,ggplot2,gridextra) #rstats https://stackoverflow.com/questions/49422380/gridextra-when-i-knit-for-pdf-i-still-get-the-big-graphs
#> 
#> [[17]]
#> How to display side by side bar plot as well as stacked bar plot in the same figure in R? (r,ggplot2,bar-chart,visualization,stacked-chart) #rstats https://stackoverflow.com/questions/49423366/how-to-display-side-by-side-bar-plot-as-well-as-stacked-bar-plot-in-the-same-fig
#> 
#> [[18]]
#> what library for Python plotting map should I use on this case scenario? (python,matplotlib,ggplot2,seaborn) #rstats https://stackoverflow.com/questions/49424518/what-library-for-python-plotting-map-should-i-use-on-this-case-scenario
#> 
#> [[19]]
#> How to arrange bars in ggplot? (r,ggplot2,dplyr) #rstats https://stackoverflow.com/questions/49426025/how-to-arrange-bars-in-ggplot
#> 
#> [[20]]
#> geom_rect alternating colour on discrete scale (r,ggplot2) #rstats https://stackoverflow.com/questions/49428048/geom-rect-alternating-colour-on-discrete-scale
#> 
#> [[21]]
#> Mistake "A continuous variable can not be mapped to share" (ggplot2) #rstats https://stackoverflow.com/questions/49429272/mistake-a-continuous-variable-can-not-be-mapped-to-share
#>