Set Up a StackOverflow Twitter Bot

Set up a StackOverflow twitter bot for a given list of tags, and excluded tags. Job scheduling can be set up using either cronR or taskscheduleR depending on your platform.

set_up_stack_tweet_bot(name = "stack_tweet_bot", extracted_tags = NULL,
  excluded_tags = NULL, time_window = 60, add_process_fn = NULL,
  hashtags = "rstats", post = TRUE, dir = NULL, run = TRUE,
  schedule = TRUE, save = TRUE, verbose = TRUE, ...)

Arguments

name
Character string containing the bots name, defaults to "stack_tweet_bot".

extracted_tags
A character vector of tags to extract questions for.

excluded_tags
A character vector of tags to exclude questions from the output.

time_window
Numeric, the timewindow to extract questions for (in minutes. Defaults to 60 minutes. Note that a maximum of 100 questions are extracted by the StackOverflow api, if you expect your tags to have a high volume of questions reduce this. If using this function as part of a sheduled bot makes sure this time frame corresponds the frequency with which your bot is running.

add_process_fn
A function. If additional munging or data extraction is required add to a custom function and pass here. This function must except a dataframe as its first arguement and return a dataframe. See the examples for details.

hashtags
A character vector of hashtags to attach to the tweet. See add_twitter_api for details.

post
Logical, defaults to TRUE. Should the formatted tweets be published on twitter, if FALSE then the tweets are instead returned as a list.

dir
Character string containing the directory into which to save the bot, defaults to the temporary directory. Is not used if save = FALSE.

run
Logical, defaults to TRUE. Should the bot be run once saved?

schedule
Logical, defaults to TRUE. Should the bot be scheduled to run using cronR/taskscheduleR? In order to schedule the bot save must be TRUE.

save
Logical, defaults to TRUE. Should the bot be saved.

verbose
Logical, defaults to TRUE. Should progress messages be printed to the terminal.

...
Additional arguements to pass to cron_add or taskscheduler_create when schedule = TRUE.The suggested frequency for the bot run is hourly, set this by passing frequency = "hourly" (cronR) or schedule = "HOURLY" (taskscheduleR).

Value

A character string containing the twitter bot code.

See also

get_stack_questions post_stack_tweets add_twitter_api add_stack_api

Examples

## Build and run twitter bot
set_up_stack_tweet_bot(extracted_tags = "ggplot2",
                       time_window = 60 * 24,
                       run = TRUE, schedule = FALSE,
                       save = TRUE, post = FALSE)

#> No directory has been supplied for saving the twitter bot,
#>            defaulting to saving to the temporary directory. This directory will not be
#>              preserved once the r session has ended.

#> Saving the following code to: /tmp/RtmpL9MihV/stack_tweet_bot.R

#> Code: 
#> 
#> library(StackTweetBot);
#> 
#> message('Getting Stack Overflow questions at ', Sys.time());
#> 
#> questions <- get_stack_questions(extracted_tags = 'ggplot2',
#>                                  excluded_tags = NULL,
#>                                  time_window = 1440,
#>                                  add_process_fn = NULL);
#> 
#> message('Extracted ', length(questions$title), ' questions');
#> 
#> message('Posting questions as tweets');
#> 
#> posts <- post_stack_tweets(questions, hashtags = 'rstats',
#>                            post = FALSE);

#> Running bot..

#> Getting Stack Overflow questions at 2018-03-22 13:59:52

#> Extracted 23 questions

#> Posting questions as tweets

#> library(StackTweetBot);
#> 
#> message('Getting Stack Overflow questions at ', Sys.time());
#> 
#> questions <- get_stack_questions(extracted_tags = 'ggplot2',
#>                                  excluded_tags = NULL,
#>                                  time_window = 1440,
#>                                  add_process_fn = NULL);
#> 
#> message('Extracted ', length(questions$title), ' questions');
#> 
#> message('Posting questions as tweets');
#> 
#> posts <- post_stack_tweets(questions, hashtags = 'rstats',
#>                            post = FALSE);

## Sample bot output
posts

#> [[1]]
#> Dynamically change background color of ggplot2 (r,ggplot2) #rstats https://stackoverflow.com/questions/49408969/dynamically-change-background-color-of-ggplot2
#> 
#> [[2]]
#> How to create multiple (6) plots with ggplot and save them to a pdf file? (r,pdf,ggplot2) #rstats https://stackoverflow.com/questions/49412035/how-to-create-multiple-6-plots-with-ggplot-and-save-them-to-a-pdf-file
#> 
#> [[3]]
#> Provide title to each of the subplots - R Shiny (r,plot,ggplot2,shiny,ggplotly) #rstats https://stackoverflow.com/questions/49412355/provide-title-to-each-of-the-subplots-r-shiny
#> 
#> [[4]]
#> How do I plot PostGIS geometries using geom_sf? (r,ggplot2,postgis,spatial) #rstats https://stackoverflow.com/questions/49413102/how-do-i-plot-postgis-geometries-using-geom-sf
#> 
#> [[5]]
#> adding legend to plot with 2 geom points (r,ggplot2) #rstats https://stackoverflow.com/questions/49415013/adding-legend-to-plot-with-2-geom-points
#> 
#> [[6]]
#> ggplot2: axis.line and panel.border have different width (r,plot,ggplot2) #rstats https://stackoverflow.com/questions/49415556/ggplot2-axis-line-and-panel-border-have-different-width
#> 
#> [[7]]
#> geom_text at certain position in a pie chart (r,ggplot2,tidyverse) #rstats https://stackoverflow.com/questions/49416386/geom-text-at-certain-position-in-a-pie-chart
#> 
#> [[8]]
#> Connected points in ggplot boxplot (r,ggplot2,data-visualization,boxplot) #rstats https://stackoverflow.com/questions/49417250/connected-points-in-ggplot-boxplot
#> 
#> [[9]]
#> R: loop through data.table and plot with ggplot2 (r,ggplot2) #rstats https://stackoverflow.com/questions/49417892/r-loop-through-data-table-and-plot-with-ggplot2
#> 
#> [[10]]
#> multivariate bar chart in R ggplot (r,ggplot2,bar-chart,data-visualization) #rstats https://stackoverflow.com/questions/49418057/multivariate-bar-chart-in-r-ggplot
#> 
#> [[11]]
#> Error in filter_impl: Evaluation error: Not compatible with requested type (r,ggplot2,shiny,dplyr) #rstats https://stackoverflow.com/questions/49418083/error-in-filter-impl-evaluation-error-not-compatible-with-requested-type
#> 
#> [[12]]
#> What is the best way to add ratio lines to plot in ggplot2? (r,plot,ggplot2,scatter-plot) #rstats https://stackoverflow.com/questions/49418545/what-is-the-best-way-to-add-ratio-lines-to-plot-in-ggplot2
#> 
#> [[13]]
#> Adding Regression Line Equation and R2 on SEPARATE LINES graph (r,ggplot2) #rstats https://stackoverflow.com/questions/49418552/adding-regression-line-equation-and-r2-on-separate-lines-graph
#> 
#> [[14]]
#> Change order of months scale_x_date ggplot (python,python-2.7,ggplot2,python-ggplot) #rstats https://stackoverflow.com/questions/49418669/change-order-of-months-scale-x-date-ggplot
#> 
#> [[15]]
#> ROC Curve in Shiny (r,ggplot2,shiny,roc,proc-r-package) #rstats https://stackoverflow.com/questions/49420225/roc-curve-in-shiny
#> 
#> [[16]]
#> Set color parameter in ggplot as optional - Shiny Dashboard (r,ggplot2,shinydashboard) #rstats https://stackoverflow.com/questions/49420853/set-color-parameter-in-ggplot-as-optional-shiny-dashboard
#> 
#> [[17]]
#> gridExtra - when I knit for pdf,i still get the big graphs (r,ggplot2,gridextra) #rstats https://stackoverflow.com/questions/49422380/gridextra-when-i-knit-for-pdf-i-still-get-the-big-graphs
#> 
#> [[18]]
#> How to display side by side bar plot as well as stacked bar plot in the same figure in R? (r,ggplot2,bar-chart,visualization,stacked-chart) #rstats https://stackoverflow.com/questions/49423366/how-to-display-side-by-side-bar-plot-as-well-as-stacked-bar-plot-in-the-same-fig
#> 
#> [[19]]
#> ggplotly function modifies line thickness and chart width (r,ggplot2,plotly,ggplotly) #rstats https://stackoverflow.com/questions/49423959/ggplotly-function-modifies-line-thickness-and-chart-width
#> 
#> [[20]]
#> what library for Python plotting map should I use on this case scenario? (python,matplotlib,ggplot2,seaborn) #rstats https://stackoverflow.com/questions/49424518/what-library-for-python-plotting-map-should-i-use-on-this-case-scenario
#> 
#> [[21]]
#> How to arrange bars in ggplot? (r,ggplot2,dplyr) #rstats https://stackoverflow.com/questions/49426025/how-to-arrange-bars-in-ggplot
#> 
#> [[22]]
#> geom_rect alternating colour on discrete scale (r,ggplot2) #rstats https://stackoverflow.com/questions/49428048/geom-rect-alternating-colour-on-discrete-scale
#> 
#> [[23]]
#> Mistake "A continuous variable can not be mapped to share" (ggplot2) #rstats https://stackoverflow.com/questions/49429272/mistake-a-continuous-variable-can-not-be-mapped-to-share
#>