Get StackOverflow Questions For A List of Tags

Extract StackOverflow questions for a given set of tags, exluding questions related to an optional list of tags. This code is based on the tidyverse tweets bot by Jake Thompson.

get_stack_questions(extracted_tags = NULL, excluded_tags = NULL,
  time_window = 60, add_process_fn = NULL, ...)

Arguments

extracted_tags
A character vector of tags to extract questions for.

excluded_tags
A character vector of tags to exclude questions from the output.

time_window
Numeric, the timewindow to extract questions for (in minutes. Defaults to 60 minutes. Note that a maximum of 100 questions are extracted by the StackOverflow api, if you expect your tags to have a high volume of questions reduce this. If using this function as part of a sheduled bot makes sure this time frame corresponds the frequency with which your bot is running.

add_process_fn
A function. If additional munging or data extraction is required add to a custom function and pass here. This function must except a dataframe as its first arguement and return a dataframe. See the examples for details.

...
Pass additional arguements to stackr::stack_questions.

Value

A list containing the title, creation date, link, and tags for StackOverflow questions related to your selected tags.

Examples

## Basic call
get_stack_questions(extracted_tags = "ggplot2",
                    excluded_tags = "plotly",
                    time_window = 60*24)

#> $title
#>  [1] "Dynamically change background color of ggplot2"                                           
#>  [2] "How to create multiple (6) plots with ggplot and save them to a pdf file?"                
#>  [3] "How do I plot PostGIS geometries using geom_sf?"                                          
#>  [4] "adding legend to plot with 2 geom points"                                                 
#>  [5] "ggplot2: axis.line and panel.border have different width"                                 
#>  [6] "geom_text at certain position in a pie chart"                                             
#>  [7] "Connected points in ggplot boxplot"                                                       
#>  [8] "R: loop through data.table and plot with ggplot2"                                         
#>  [9] "multivariate bar chart in R ggplot"                                                       
#> [10] "Error in filter_impl: Evaluation error: Not compatible with requested type"               
#> [11] "What is the best way to add ratio lines to plot in ggplot2?"                              
#> [12] "Adding Regression Line Equation and R2 on SEPARATE LINES graph"                           
#> [13] "Change order of months scale_x_date ggplot"                                               
#> [14] "ROC Curve in Shiny"                                                                       
#> [15] "Set color parameter in ggplot as optional - Shiny Dashboard"                              
#> [16] "gridExtra - when I knit for pdf,i still get the big graphs"                               
#> [17] "How to display side by side bar plot as well as stacked bar plot in the same figure in R?"
#> [18] "what library for Python plotting map should I use on this case scenario?"                 
#> [19] "How to arrange bars in ggplot?"                                                           
#> [20] "geom_rect alternating colour on discrete scale"                                           
#> [21] "Mistake \"A continuous variable can not be mapped to share\""                             
#> 
#> $creation_date
#>  [1] "2018-03-21 14:27:59 UTC" "2018-03-21 16:42:06 UTC"
#>  [3] "2018-03-21 17:36:07 UTC" "2018-03-21 19:26:04 UTC"
#>  [5] "2018-03-21 20:01:59 UTC" "2018-03-21 20:58:02 UTC"
#>  [7] "2018-03-21 22:02:10 UTC" "2018-03-21 23:01:09 UTC"
#>  [9] "2018-03-21 23:18:43 UTC" "2018-03-21 23:21:21 UTC"
#> [11] "2018-03-22 00:14:56 UTC" "2018-03-22 00:15:29 UTC"
#> [13] "2018-03-22 00:30:05 UTC" "2018-03-22 03:51:52 UTC"
#> [15] "2018-03-22 04:58:27 UTC" "2018-03-22 06:57:17 UTC"
#> [17] "2018-03-22 08:00:29 UTC" "2018-03-22 09:07:34 UTC"
#> [19] "2018-03-22 10:18:02 UTC" "2018-03-22 11:54:45 UTC"
#> [21] "2018-03-22 12:55:08 UTC"
#> 
#> $link
#>  [1] "https://stackoverflow.com/questions/49408969/dynamically-change-background-color-of-ggplot2"                                  
#>  [2] "https://stackoverflow.com/questions/49412035/how-to-create-multiple-6-plots-with-ggplot-and-save-them-to-a-pdf-file"          
#>  [3] "https://stackoverflow.com/questions/49413102/how-do-i-plot-postgis-geometries-using-geom-sf"                                  
#>  [4] "https://stackoverflow.com/questions/49415013/adding-legend-to-plot-with-2-geom-points"                                        
#>  [5] "https://stackoverflow.com/questions/49415556/ggplot2-axis-line-and-panel-border-have-different-width"                         
#>  [6] "https://stackoverflow.com/questions/49416386/geom-text-at-certain-position-in-a-pie-chart"                                    
#>  [7] "https://stackoverflow.com/questions/49417250/connected-points-in-ggplot-boxplot"                                              
#>  [8] "https://stackoverflow.com/questions/49417892/r-loop-through-data-table-and-plot-with-ggplot2"                                 
#>  [9] "https://stackoverflow.com/questions/49418057/multivariate-bar-chart-in-r-ggplot"                                              
#> [10] "https://stackoverflow.com/questions/49418083/error-in-filter-impl-evaluation-error-not-compatible-with-requested-type"        
#> [11] "https://stackoverflow.com/questions/49418545/what-is-the-best-way-to-add-ratio-lines-to-plot-in-ggplot2"                      
#> [12] "https://stackoverflow.com/questions/49418552/adding-regression-line-equation-and-r2-on-separate-lines-graph"                  
#> [13] "https://stackoverflow.com/questions/49418669/change-order-of-months-scale-x-date-ggplot"                                      
#> [14] "https://stackoverflow.com/questions/49420225/roc-curve-in-shiny"                                                              
#> [15] "https://stackoverflow.com/questions/49420853/set-color-parameter-in-ggplot-as-optional-shiny-dashboard"                       
#> [16] "https://stackoverflow.com/questions/49422380/gridextra-when-i-knit-for-pdf-i-still-get-the-big-graphs"                        
#> [17] "https://stackoverflow.com/questions/49423366/how-to-display-side-by-side-bar-plot-as-well-as-stacked-bar-plot-in-the-same-fig"
#> [18] "https://stackoverflow.com/questions/49424518/what-library-for-python-plotting-map-should-i-use-on-this-case-scenario"         
#> [19] "https://stackoverflow.com/questions/49426025/how-to-arrange-bars-in-ggplot"                                                   
#> [20] "https://stackoverflow.com/questions/49428048/geom-rect-alternating-colour-on-discrete-scale"                                  
#> [21] "https://stackoverflow.com/questions/49429272/mistake-a-continuous-variable-can-not-be-mapped-to-share"                        
#> 
#> $tags
#>  [1] "r,ggplot2"                                      
#>  [2] "r,pdf,ggplot2"                                  
#>  [3] "r,ggplot2,postgis,spatial"                      
#>  [4] "r,ggplot2"                                      
#>  [5] "r,plot,ggplot2"                                 
#>  [6] "r,ggplot2,tidyverse"                            
#>  [7] "r,ggplot2,data-visualization,boxplot"           
#>  [8] "r,ggplot2"                                      
#>  [9] "r,ggplot2,bar-chart,data-visualization"         
#> [10] "r,ggplot2,shiny,dplyr"                          
#> [11] "r,plot,ggplot2,scatter-plot"                    
#> [12] "r,ggplot2"                                      
#> [13] "python,python-2.7,ggplot2,python-ggplot"        
#> [14] "r,ggplot2,shiny,roc,proc-r-package"             
#> [15] "r,ggplot2,shinydashboard"                       
#> [16] "r,ggplot2,gridextra"                            
#> [17] "r,ggplot2,bar-chart,visualization,stacked-chart"
#> [18] "python,matplotlib,ggplot2,seaborn"              
#> [19] "r,ggplot2,dplyr"                                
#> [20] "r,ggplot2"                                      
#> [21] "ggplot2"                                        
#> 

## Add additional processing step
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

add_process_fn <- function(df) {
      df %>%
      mutate(id = 1:n())
}

get_stack_questions(extracted_tags = "ggplot2",
                    excluded_tags = "plotly",
                    time_window = 60*24,
                    add_process_fn = add_process_fn)

#> $title
#>  [1] "Dynamically change background color of ggplot2"                                           
#>  [2] "How to create multiple (6) plots with ggplot and save them to a pdf file?"                
#>  [3] "How do I plot PostGIS geometries using geom_sf?"                                          
#>  [4] "adding legend to plot with 2 geom points"                                                 
#>  [5] "ggplot2: axis.line and panel.border have different width"                                 
#>  [6] "geom_text at certain position in a pie chart"                                             
#>  [7] "Connected points in ggplot boxplot"                                                       
#>  [8] "R: loop through data.table and plot with ggplot2"                                         
#>  [9] "multivariate bar chart in R ggplot"                                                       
#> [10] "Error in filter_impl: Evaluation error: Not compatible with requested type"               
#> [11] "What is the best way to add ratio lines to plot in ggplot2?"                              
#> [12] "Adding Regression Line Equation and R2 on SEPARATE LINES graph"                           
#> [13] "Change order of months scale_x_date ggplot"                                               
#> [14] "ROC Curve in Shiny"                                                                       
#> [15] "Set color parameter in ggplot as optional - Shiny Dashboard"                              
#> [16] "gridExtra - when I knit for pdf,i still get the big graphs"                               
#> [17] "How to display side by side bar plot as well as stacked bar plot in the same figure in R?"
#> [18] "what library for Python plotting map should I use on this case scenario?"                 
#> [19] "How to arrange bars in ggplot?"                                                           
#> [20] "geom_rect alternating colour on discrete scale"                                           
#> [21] "Mistake \"A continuous variable can not be mapped to share\""                             
#> 
#> $creation_date
#>  [1] "2018-03-21 14:27:59 UTC" "2018-03-21 16:42:06 UTC"
#>  [3] "2018-03-21 17:36:07 UTC" "2018-03-21 19:26:04 UTC"
#>  [5] "2018-03-21 20:01:59 UTC" "2018-03-21 20:58:02 UTC"
#>  [7] "2018-03-21 22:02:10 UTC" "2018-03-21 23:01:09 UTC"
#>  [9] "2018-03-21 23:18:43 UTC" "2018-03-21 23:21:21 UTC"
#> [11] "2018-03-22 00:14:56 UTC" "2018-03-22 00:15:29 UTC"
#> [13] "2018-03-22 00:30:05 UTC" "2018-03-22 03:51:52 UTC"
#> [15] "2018-03-22 04:58:27 UTC" "2018-03-22 06:57:17 UTC"
#> [17] "2018-03-22 08:00:29 UTC" "2018-03-22 09:07:34 UTC"
#> [19] "2018-03-22 10:18:02 UTC" "2018-03-22 11:54:45 UTC"
#> [21] "2018-03-22 12:55:08 UTC"
#> 
#> $link
#>  [1] "https://stackoverflow.com/questions/49408969/dynamically-change-background-color-of-ggplot2"                                  
#>  [2] "https://stackoverflow.com/questions/49412035/how-to-create-multiple-6-plots-with-ggplot-and-save-them-to-a-pdf-file"          
#>  [3] "https://stackoverflow.com/questions/49413102/how-do-i-plot-postgis-geometries-using-geom-sf"                                  
#>  [4] "https://stackoverflow.com/questions/49415013/adding-legend-to-plot-with-2-geom-points"                                        
#>  [5] "https://stackoverflow.com/questions/49415556/ggplot2-axis-line-and-panel-border-have-different-width"                         
#>  [6] "https://stackoverflow.com/questions/49416386/geom-text-at-certain-position-in-a-pie-chart"                                    
#>  [7] "https://stackoverflow.com/questions/49417250/connected-points-in-ggplot-boxplot"                                              
#>  [8] "https://stackoverflow.com/questions/49417892/r-loop-through-data-table-and-plot-with-ggplot2"                                 
#>  [9] "https://stackoverflow.com/questions/49418057/multivariate-bar-chart-in-r-ggplot"                                              
#> [10] "https://stackoverflow.com/questions/49418083/error-in-filter-impl-evaluation-error-not-compatible-with-requested-type"        
#> [11] "https://stackoverflow.com/questions/49418545/what-is-the-best-way-to-add-ratio-lines-to-plot-in-ggplot2"                      
#> [12] "https://stackoverflow.com/questions/49418552/adding-regression-line-equation-and-r2-on-separate-lines-graph"                  
#> [13] "https://stackoverflow.com/questions/49418669/change-order-of-months-scale-x-date-ggplot"                                      
#> [14] "https://stackoverflow.com/questions/49420225/roc-curve-in-shiny"                                                              
#> [15] "https://stackoverflow.com/questions/49420853/set-color-parameter-in-ggplot-as-optional-shiny-dashboard"                       
#> [16] "https://stackoverflow.com/questions/49422380/gridextra-when-i-knit-for-pdf-i-still-get-the-big-graphs"                        
#> [17] "https://stackoverflow.com/questions/49423366/how-to-display-side-by-side-bar-plot-as-well-as-stacked-bar-plot-in-the-same-fig"
#> [18] "https://stackoverflow.com/questions/49424518/what-library-for-python-plotting-map-should-i-use-on-this-case-scenario"         
#> [19] "https://stackoverflow.com/questions/49426025/how-to-arrange-bars-in-ggplot"                                                   
#> [20] "https://stackoverflow.com/questions/49428048/geom-rect-alternating-colour-on-discrete-scale"                                  
#> [21] "https://stackoverflow.com/questions/49429272/mistake-a-continuous-variable-can-not-be-mapped-to-share"                        
#> 
#> $tags
#>  [1] "r,ggplot2"                                      
#>  [2] "r,pdf,ggplot2"                                  
#>  [3] "r,ggplot2,postgis,spatial"                      
#>  [4] "r,ggplot2"                                      
#>  [5] "r,plot,ggplot2"                                 
#>  [6] "r,ggplot2,tidyverse"                            
#>  [7] "r,ggplot2,data-visualization,boxplot"           
#>  [8] "r,ggplot2"                                      
#>  [9] "r,ggplot2,bar-chart,data-visualization"         
#> [10] "r,ggplot2,shiny,dplyr"                          
#> [11] "r,plot,ggplot2,scatter-plot"                    
#> [12] "r,ggplot2"                                      
#> [13] "python,python-2.7,ggplot2,python-ggplot"        
#> [14] "r,ggplot2,shiny,roc,proc-r-package"             
#> [15] "r,ggplot2,shinydashboard"                       
#> [16] "r,ggplot2,gridextra"                            
#> [17] "r,ggplot2,bar-chart,visualization,stacked-chart"
#> [18] "python,matplotlib,ggplot2,seaborn"              
#> [19] "r,ggplot2,dplyr"                                
#> [20] "r,ggplot2"                                      
#> [21] "ggplot2"                                        
#> 
#> $id
#>  [1] 21 20 19 18 17 16 15 14 13 12 11 10  9  8  7  6  5  4  3  2  1
#>