Get StackOverflow Questions For A List of Tags

Extract StackOverflow questions for a given set of tags, exluding questions related to an optional list of tags. This code is based on the tidyverse tweets bot by Jake Thompson.

get_stack_questions(extracted_tags = NULL, excluded_tags = NULL,
  time_window = 60, Rprofile_path = "~/.Rprofile", add_process_fn = NULL)

Arguments

extracted_tags
A character vector of tags to extract questions for.

excluded_tags
A character vector of tags to exclude questions from the output.

time_window
Numeric, the timewindow to extract questions for (in minutes. Defaults to 60 minutes. Note that a maximum of 100 questions are extracted by the StackOverflow api, if you expect your tags to have a high volume of questions reduce this. If using this function as part of a sheduled bot makes sure this time frame corresponds the frequency with which your bot is running.

Rprofile_path
A character string containing the path to your .Rprofile. Defaults to your home directory.

add_process_fn
A function. If additional munging or data extraction is required add to a custom function and pass here. This function must except a dataframe as its first arguement and return a dataframe. See the examples for details.

Value

A list containing the title, creation date, link, and tags for StackOverflow questions related to your selected tags.

Examples

## Basic call
get_stack_questions(extracted_tags = "ggplot2",
                    excluded_tags = "plotly",
                    time_window = 120)

#> $title
#> [1] "Getting \"npc\" coordinates of ggplot2 grob"  
#> [2] "Treemap colors mapped to category in R"       
#> [3] "Grouped Barchart y-Axis -- ggplot"            
#> [4] "ggplot2::geom_tile can not assign fill color?"
#> [5] "Create separated US regions"                  
#> 
#> $creation_date
#> [1] "2018-02-09 16:47:23 GMT" "2018-02-09 17:16:46 GMT"
#> [3] "2018-02-09 17:42:44 GMT" "2018-02-09 17:45:24 GMT"
#> [5] "2018-02-09 18:21:24 GMT"
#> 
#> $link
#> [1] "https://stackoverflow.com/questions/48710478/getting-npc-coordinates-of-ggplot2-grob"   
#> [2] "https://stackoverflow.com/questions/48710961/treemap-colors-mapped-to-category-in-r"    
#> [3] "https://stackoverflow.com/questions/48711358/grouped-barchart-y-axis-ggplot"            
#> [4] "https://stackoverflow.com/questions/48711408/ggplot2geom-tile-can-not-assign-fill-color"
#> [5] "https://stackoverflow.com/questions/48711914/create-separated-us-regions"               
#> 
#> $tags
#> [1] "r,ggplot2,lattice,gtable" "r,ggplot2,treemap"       
#> [3] "r,ggplot2"                "r,ggplot2,fill"          
#> [5] "r,ggplot2,maps,census"   
#> 

## Add additional processing step
library(dplyr)
add_process_fn <- function(df) {
      df %>%
      mutate(id = 1:n())
}

get_stack_questions(extracted_tags = "ggplot2",
                    excluded_tags = "plotly",
                    time_window = 120,
                    add_process_fn = add_process_fn)

#> $title
#> [1] "Getting \"npc\" coordinates of ggplot2 grob"  
#> [2] "Treemap colors mapped to category in R"       
#> [3] "Grouped Barchart y-Axis -- ggplot"            
#> [4] "ggplot2::geom_tile can not assign fill color?"
#> [5] "Create separated US regions"                  
#> 
#> $creation_date
#> [1] "2018-02-09 16:47:23 GMT" "2018-02-09 17:16:46 GMT"
#> [3] "2018-02-09 17:42:44 GMT" "2018-02-09 17:45:24 GMT"
#> [5] "2018-02-09 18:21:24 GMT"
#> 
#> $link
#> [1] "https://stackoverflow.com/questions/48710478/getting-npc-coordinates-of-ggplot2-grob"   
#> [2] "https://stackoverflow.com/questions/48710961/treemap-colors-mapped-to-category-in-r"    
#> [3] "https://stackoverflow.com/questions/48711358/grouped-barchart-y-axis-ggplot"            
#> [4] "https://stackoverflow.com/questions/48711408/ggplot2geom-tile-can-not-assign-fill-color"
#> [5] "https://stackoverflow.com/questions/48711914/create-separated-us-regions"               
#> 
#> $tags
#> [1] "r,ggplot2,lattice,gtable" "r,ggplot2,treemap"       
#> [3] "r,ggplot2"                "r,ggplot2,fill"          
#> [5] "r,ggplot2,maps,census"   
#> 
#> $id
#> [1] 5 4 3 2 1
#>