Post A Tweet for Each Question

Format and post StackOverflow questions as tweets, optionally returning them as a formatted list. This code is based on the tidyverse tweets bot by Jake Thompson.

post_stack_tweets(questions = NULL, hashtags = "rstats", post = TRUE)

Arguments

questions
A list of questions to tweet as produced by get_stack_questions.

hashtags
A character vector of hashtags to attach to the tweet. See add_twitter_api for details.

post
Logical, defaults to TRUE. Should the formatted tweets be published on twitter, if FALSE then the tweets are instead returned as a list.

Value

Either nothing is returned or a formated list of tweets.

Examples

ggplot2_qs <-  get_stack_questions(extracted_tags = "ggplot2",
                                  excluded_tags = "plotly",
                                  time_window = 60*24)

post_stack_tweets(ggplot2_qs, hashtags = "rstats", post = FALSE)

#> [[1]]
#> ggplot2 - Correctly arrange odd number of plots into one figure (r,ggplot2,cowplot) #rstats https://stackoverflow.com/questions/52953569/ggplot2-correctly-arrange-odd-number-of-plots-into-one-figure
#> 
#> [[2]]
#> Plotting monthly percentages in shiny R (r,ggplot2,shiny,shinydashboard) #rstats https://stackoverflow.com/questions/52954681/plotting-monthly-percentages-in-shiny-r
#> 
#> [[3]]
#> ggplot order of geom_point and geom_line differ (r,ggplot2) #rstats https://stackoverflow.com/questions/52954846/ggplot-order-of-geom-point-and-geom-line-differ
#> 
#> [[4]]
#> How do I annotate a ggtree object with bootstrap values? (r,ggplot2,bioinformatics,phylogeny,ggtree) #rstats https://stackoverflow.com/questions/52956181/how-do-i-annotate-a-ggtree-object-with-bootstrap-values
#> 
#> [[5]]
#> How to annotate different values for each facet with dodged geom_boxplot on R? (r,ggplot2,text,boxplot) #rstats https://stackoverflow.com/questions/52958453/how-to-annotate-different-values-for-each-facet-with-dodged-geom-boxplot-on-r
#> 
#> [[6]]
#> How to fix the multiple error bar issue of ggplot in R? (ggplot2,errorbar) #rstats https://stackoverflow.com/questions/52960261/how-to-fix-the-multiple-error-bar-issue-of-ggplot-in-r
#> 
#> [[7]]
#> No legend in ggplot line graph (r,ggplot2,legend,linegraph) #rstats https://stackoverflow.com/questions/52960885/no-legend-in-ggplot-line-graph
#> 
#> [[8]]
#> Choropleth Plotting polygons with ggplot2 R on a map (r,ggplot2,gis,ggmap) #rstats https://stackoverflow.com/questions/52961807/choropleth-plotting-polygons-with-ggplot2-r-on-a-map
#> 
#> [[9]]
#> Why do I get the weird purple line in ggplot? (r,ggplot2) #rstats https://stackoverflow.com/questions/52961934/why-do-i-get-the-weird-purple-line-in-ggplot
#> 
#> [[10]]
#> R (ggplot2): line with data labels (r,ggplot2) #rstats https://stackoverflow.com/questions/52962517/r-ggplot2-line-with-data-labels
#> 
#> [[11]]
#> ggplot y axis adjustment and statistics (r,ggplot2,ggpubr) #rstats https://stackoverflow.com/questions/52966608/ggplot-y-axis-adjustment-and-statistics
#> 
#> [[12]]
#> How to install ggplot2 for R to a custom lib directory (r,ggplot2,installation,package) #rstats https://stackoverflow.com/questions/52967249/how-to-install-ggplot2-for-r-to-a-custom-lib-directory
#> 
#> [[13]]
#> ggplot2: can geom_point work in melted dataframes? (r,ggplot2,scatter-plot,dcast) #rstats https://stackoverflow.com/questions/52968996/ggplot2-can-geom-point-work-in-melted-dataframes
#> 
#> [[14]]
#> Framework for adding lines on existing plots: "points" and "plot" in ggplot (r,ggplot2) #rstats https://stackoverflow.com/questions/52970305/framework-for-adding-lines-on-existing-plots-points-and-plot-in-ggplot
#> 
#> [[15]]
#> grouped barplot from table (r,ggplot2,bar-chart) #rstats https://stackoverflow.com/questions/52970306/grouped-barplot-from-table
#> 
#> [[16]]
#> ggplot: Show 24-hour activity profiles for each day through 48 weeks (r,ggplot2,plot) #rstats https://stackoverflow.com/questions/52972439/ggplot-show-24-hour-activity-profiles-for-each-day-through-48-weeks
#> 
#> [[17]]
#> ggplot2: grouping a barplot by several instead of a single categorical variable (r,ggplot2) #rstats https://stackoverflow.com/questions/52973023/ggplot2-grouping-a-barplot-by-several-instead-of-a-single-categorical-variable
#> 
#> [[18]]
#> Plotting multiple densities in ggplot2: How to add an extra variable with a different scale? (r,ggplot2,scaling) #rstats https://stackoverflow.com/questions/52973132/plotting-multiple-densities-in-ggplot2-how-to-add-an-extra-variable-with-a-diff
#> 
#> [[19]]
#> How to pass a string argument to expression function in ggplot2? (ggplot2) #rstats https://stackoverflow.com/questions/52973186/how-to-pass-a-string-argument-to-expression-function-in-ggplot2
#>