Post A Tweet for Each Question

Format and post StackOverflow questions as tweets, optionally returning them as a formatted list. This code is based on the tidyverse tweets bot by Jake Thompson.

post_stack_tweets(questions = NULL, hashtags = "rstats", post = TRUE)

Arguments

questions
A list of questions to tweet as produced by get_stack_questions.

hashtags
A character vector of hashtags to attach to the tweet. See add_twitter_api for details.

post
Logical, defaults to TRUE. Should the formatted tweets be published on twitter, if FALSE then the tweets are instead returned as a list.

Value

Either nothing is returned or a formated list of tweets.

Examples

ggplot2_qs <-  get_stack_questions(extracted_tags = "ggplot2",
                                  excluded_tags = "plotly",
                                  time_window = 120)

post_stack_tweets(ggplot2_qs, hashtags = "rstats", post = FALSE)

#> [[1]]
#> Countinuous outline in stacked ggplot2 barplot (r,ggplot2) #rstats https://stackoverflow.com/questions/49223425/countinuous-outline-in-stacked-ggplot2-barplot
#> 
#> [[2]]
#> ggplot2 scale_fill_grey for boxplot (r,ggplot2) #rstats https://stackoverflow.com/questions/49223476/ggplot2-scale-fill-grey-for-boxplot
#>