Set Up a StackOverflow Twitter Bot

Set up a StackOverflow twitter bot for a given list of tags, and excluded tags. Job scheduling can be set up using either cronR or taskscheduleR depending on your platform.

set_up_stack_tweet_bot(name = "stack_tweet_bot", extracted_tags = NULL,
  excluded_tags = NULL, time_window = 60, add_process_fn = NULL,
  hashtags = "rstats", post = TRUE, dir = NULL, run = TRUE,
  schedule = TRUE, save = TRUE, verbose = TRUE, ...)

Arguments

name
Character string containing the bots name, defaults to "stack_tweet_bot".

extracted_tags
A character vector of tags to extract questions for.

excluded_tags
A character vector of tags to exclude questions from the output.

time_window
Numeric, the timewindow to extract questions for (in minutes. Defaults to 60 minutes. Note that a maximum of 100 questions are extracted by the StackOverflow api, if you expect your tags to have a high volume of questions reduce this. If using this function as part of a sheduled bot makes sure this time frame corresponds the frequency with which your bot is running.

add_process_fn
A function. If additional munging or data extraction is required add to a custom function and pass here. This function must except a dataframe as its first arguement and return a dataframe. See the examples for details.

hashtags
A character vector of hashtags to attach to the tweet. See add_twitter_api for details.

post
Logical, defaults to TRUE. Should the formatted tweets be published on twitter, if FALSE then the tweets are instead returned as a list.

dir
Character string containing the directory into which to save the bot, defaults to the temporary directory. Is not used if save = FALSE.

run
Logical, defaults to TRUE. Should the bot be run once saved?

schedule
Logical, defaults to TRUE. Should the bot be scheduled to run using cronR/taskscheduleR? In order to schedule the bot save must be TRUE.

save
Logical, defaults to TRUE. Should the bot be saved.

verbose
Logical, defaults to TRUE. Should progress messages be printed to the terminal.

...
Additional arguements to pass to cron_add or taskscheduler_create when schedule = TRUE.The suggested frequency for the bot run is hourly, set this by passing frequency = "hourly" (cronR) or schedule = "HOURLY" (taskscheduleR).

Value

A character string containing the twitter bot code.

See also

get_stack_questions post_stack_tweets add_twitter_api add_stack_exchange_api

Examples

## Build and run twitter bot
set_up_stack_tweet_bot(extracted_tags = "ggplot2",
                       run = TRUE, schedule = FALSE,
                       save = TRUE, post = FALSE)

#> No directory has been supplied for saving the twitter bot,
#>            defaulting to saving to the temporary directory. This directory will not be
#>              preserved once the r session has ended.

#> Saving the following code to: /tmp/Rtmpj1Wdef/stack_tweet_bot.R

#> Code: 
#> 
#> library(StackTweetBot);
#> questions <- get_stack_questions(extracted_tags = 'ggplot2',
#>                                  excluded_tags = NULL,
#>                                  time_window = 60,
#>                                  add_process_fn = NULL);
#> 
#> posts <- post_stack_tweets(questions, hashtags = 'rstats',
#>                            post = FALSE);

#> Running bot..

#> library(StackTweetBot);
#> questions <- get_stack_questions(extracted_tags = 'ggplot2',
#>                                  excluded_tags = NULL,
#>                                  time_window = 60,
#>                                  add_process_fn = NULL);
#> 
#> posts <- post_stack_tweets(questions, hashtags = 'rstats',
#>                            post = FALSE);

## Sample bot output
posts

#> [[1]]
#> Legend does not fit well ggplot density (ggplot2,label,legend) #rstats https://stackoverflow.com/questions/49210823/legend-does-not-fit-well-ggplot-density
#> 
#> [[2]]
#> Plot proportions of dummy over time for three groups (r,plot,ggplot2) #rstats https://stackoverflow.com/questions/49211011/plot-proportions-of-dummy-over-time-for-three-groups
#>